Changed method of installing Ansible Roles
No longer using ansible-galaxy install -r requirements.yml as you are not able to use branches
to checkout. You can use tags but not wanting to do that at this time. So we are now cloning
the GitHub repo for the Ansible role and then checking out the correct branch prior to running
the playbook.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>